feat: amélioration des requetes pour l'application Android, le debut d'integration des tokens JWT pour l'application Android.
