+++ server/routes/react.js
@@ -1,15 +1,7 @@
 import {Router} from 'express';
-import {
-    activateCard,
-    addCard,
-    getCardData,
-    getNewAccessToken,
-    login,
-    logout,
-    signup, updateProfile,
-    verifyBalance
-} from '../controllers/react.js';
 import {authenticateJWT} from "../middleware/authMiddleware.js";
+import {getNewAccessToken, login, logout, signup, updateProfile} from "../controllers/react/users/connection.js";
+import {activateCard, addCard, getCardData} from "../controllers/react/users/gestion_carte.js";
 
 const routerReact = Router();
 
@@ -22,5 +14,5 @@ routerReact.post('/updateProfile', authenticateJWT, updateProfile);
 routerReact.post('/addCard', authenticateJWT, addCard);
 routerReact.post('/activateCard', authenticateJWT, activateCard);
 routerReact.get('/getCardData', authenticateJWT, getCardData);
-routerReact.get('/verifyBalance', authenticateJWT, verifyBalance);
+
 export default routerReact;
